refactor(satellite): group auth assertions into a module
