+++ src/libs/satellite/src/auth/assert/mod.rs
@@ -1,0 +1,2 @@
+pub mod caller;
+pub mod config;
