+++ src/libs/satellite/src/auth/mod.rs
@@ -1,6 +1,7 @@
 mod alternative_origins;
 mod assert;
-pub mod assert_caller;
 mod state;
 pub mod store;
 pub mod types;
+
+pub use assert::caller;
