+++ src/libs/satellite/src/auth/store.rs
@@ -1,5 +1,5 @@
 use crate::auth::alternative_origins::update_alternative_origins;
-use crate::auth::assert::assert_config_origin_urls;
+use crate::auth::assert::config::assert_config_origin_urls;
 use crate::auth::state::{get_config as get_state_config, insert_config as insert_state_config};
 use crate::auth::types::config::AuthenticationConfig;
